+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+## Changed
+
+- Run extension-less tools such as `leanc` using `sh` on Windows (and hope for the best...)
+
 # 1.0.4 - 2021-05-24
 
 ## Changed